requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
attrs==23.2.0
autoflake==2.3.1
bcrypt==4.0.0
billiard==4.2.0
black==24.4.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.0
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==7.0.0
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.6
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.15
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
pyright==1.1.358
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
sentry-sdk==1.45.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==2.0.28
starlette==0.36.3
time-machine==2.14.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
WTForms==3.1.2
yarl==1.9.4
